chore: move county logic in `useGetCounties` and delete user context
