--- react-app/src/components/Context/index.ts
@@ -1,1 +1,0 @@
-export * from "./userContext";
